pyproject.toml
[project]
name = "exam-question-downloader"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "serpucga", email = "serpucga@inf.upv.es" }
]
dependencies = [
    "jupyterlab>=4.1.8",
    "selenium>=4.20.0",
    "requests>=2.31.0",
    "logutils-serpucga>=0.1.9",
    "pymupdf>=1.24.2",
    "pytesseract>=0.3.10",
    "pandas>=2.2.2",
    "pdf2image>=1.17.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/exam_question_downloader"]
